pyproject.toml
[tool.poetry]
name = "encoding-api-using-sentence-transformers"
version = "0.1.0"
description = ""
authors = ["Shogo Ogami <sg.tao.so@gmail.com>"]
license = "MIT-License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
sentence-transformers = "^2.7.0"
waitress = "^3.0.0"
faiss-cpu = "^1.8.0"
numpy = "^1.26.4"
pandas = "^2.2.2"


[tool.poetry.group.dev.dependencies]
numpy = "^1.26.4"
pandas = "^2.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
